refactor: split editorial_generator.py into focused renderer modules (Phase 3)
Extract two large HTML-rendering methods into dedicated modules:

- editorial_renderer.py (824 lines): generate_article_html() and
  generate_amp_html() — standalone functions for individual article pages
  and their AMP counterparts. Also hosts _safe_href() (moved from
  editorial_generator).

- articles_index_renderer.py (1128 lines): generate_articles_index_html()
  — the full /articles/ index page with search, filter, sort, pagination,
  and keyboard navigation.

editorial_generator.py drops from 3330 → 1448 lines. All three fat methods
(_generate_article_html 553 lines, _generate_amp_html 230 lines,
generate_articles_index 1135 lines) are now thin 1-line wrappers that
delegate to the new modules.
